Consolidate review tests into test_patch_gemma4_text_mlp.py
